Move open-id-config.js to aspsp-open-id-config/index.js
Move test/openid-test.js to test/aspsp-open-id-config/index-test.js
