+++ lib/app.js
@@ -2,7 +2,7 @@ const express = require('express');
 const { OBAccountPaymentServiceProviders } = require('./ob-directory');
 const { authServer } = require('./aspsp-authorisation-server/token');
 const { accountRequests } = require('./aspsp-authorisation-server/account-requests');
-const { openIdConfig } = require('./open-id-config');
+const { openIdConfig } = require('./aspsp-open-id-config');
 const bodyParser = require('body-parser');
 
 const app = express();
